+++ Utilities/DumpData/DumpData.cs
@@ -25,7 +25,35 @@ namespace Sheepy.PhoenixPt.DumpData {
       public bool   Dump_Command_Csv = true;
       public bool   Dump_Guid_Csv = true;
       public bool   Dump_Lang_Csv = true;
-      public string[] Dump_Defs = new string[]{ "EntitlementDef", "VehicleItemDef", "TacticalItemDef", "TacticalActorDef", "TacUnitClassDef", "TacMissionTypeDef", "SpecializationDef", "ResearchDef", "PhoenixFacilityDef", "GroundVehicleItemDef", "GeoscapeEventDef", "GeoSiteSceneDef", "GeoMistGeneratorDef", "GeoHavenZoneDef", "GeoFactionDef", "GeoAlienBaseDef", "GeoActorDef", "GameTagDef", "DamageKeywordDef", "ComponentSetDef", "BodyPartAspectDef", "AchievementDef", "AbilityTrackDef", "AbilityDef" };
+      public string[] Dump_Defs = new string[]{
+         "VehicleItemDef",
+         "TacticalItemDef",
+         "TacticalActorDef",
+         "TacUnitClassDef",
+         "TacMissionTypeDef",
+         "SpecializationDef",
+         "SharedGameTagsDataDef",
+         "ResearchDef",
+         "PhoenixFacilityDef",
+         "HumanCustomizationDef",
+         "GroundVehicleItemDef",
+         "GeoscapeEventDef",
+         "GeoSiteSceneDef",
+         "GeoMistGeneratorDef",
+         "GeoHavenZoneDef",
+         "GeoFactionDef",
+         "GeoAlienBaseDef",
+         "GeoActorDef",
+         "GameTagDef",
+         "EntitlementDef",
+         "DamageKeywordDef",
+         "CustomizationTagDef",
+         "ComponentSetDef",
+         "BodyPartAspectDef",
+         "AchievementDef",
+         "AbilityTrackDef",
+         "AbilityDef",
+      };
       public string[] Dump_Others = new string[] {
          "GameSettings", "SharedData.GetSharedDataFromGame().AISettingsDef",
          "GameSettings", "SharedData.GetSharedDataFromGame().ContributionSettings",
@@ -33,14 +61,15 @@ namespace Sheepy.PhoenixPt.DumpData {
          "GameSettings", "SharedData.GetSharedDataFromGame().DiplomacySettings",
          "GameSettings", "SharedData.GetSharedDataFromGame().DynamicDifficultySettings",
       };
-      public uint   Config_Version = 20201128;
+      public uint   Config_Version = 20210129;
 
       internal void Upgrade () {
          if ( Config_Version >= 20201128 ) return;
          //if ( Config_Version < 20200630 ) ; // Do nothing; Added Auto_Dump field
-         if ( Config_Version < 20201128 )
+         //if ( Config_Version < 20201128 ) ; // Never released and merged with 20210129. Added EntitlementDef.
+         if ( Config_Version < 20210129 )
             if ( Dump_Defs?.Length == 23 && Dump_Defs[0] == "VehicleItemDef" && Dump_Defs[22] == "AbilityDef" )
-               Dump_Defs = new ModConfig().Dump_Defs; // Add EntitlementDef
+               Dump_Defs = new ModConfig().Dump_Defs; // Add CustomizationTagDef, EntitlementDef, HumanCustomizationDef, SharedGameTagsDataDef
          ZyMod.Api( "config save", this ); // Add Auto_Dump
       }
    }
@@ -175,13 +204,13 @@ namespace Sheepy.PhoenixPt.DumpData {
       private static void AddOthersToDump () {
          var list = Config.Dump_Others;
          if ( list == null || list.Length == 0 ) return;
-         if ( Api( "api_info", "eval.cs" ) == null ) {
-            Warn( "'Dump_Others' requires Scripting Library or other mod that provides the 'eval.cs' api.  Delete it from config to remove this warning." );
+         if ( Api( "api_info", "eval.js" ) == null ) {
+            Warn( "'Dump_Others' requires Scripting Library or other mod that provides the 'eval.js' api.  Delete it from config to remove this warning." );
             return;
          }
          for ( var i = 0 ; i < list.Length-1 ; i += 2 ) {
             string name = list[ i ] ?? "Others", cmd = list[ i + 1 ];
-            var data = Api( "eval.cs", cmd );
+            var data = Api( "eval.js", cmd );
             if ( data == null ) {
                Info( "{0} is null", cmd );
                continue;